Relocated interface from prompt file
